+++ mlxp/errors.py
@@ -30,3 +30,12 @@ class InvalidAggregationMapError(Exception):
     """Raised when an aggregation map is not an instance of AggregationMap."""
 
     pass
+
+class InvalidShellPathError(Exception):
+    """Raised when no shell path is specified"""
+    pass
+
+class UnknownSystemError(Exception):
+    """Raised when system is different from Linux, Darwin or Windows"""
+    pass 
+