+++ project_info.py
@@ -7,5 +7,5 @@ LICENSE = "MIT License"
 COPYRIGHT = "Copyright (C) 2023 " + AUTHOR
 
 
-
+