+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 # Changes
 
-- Fix compatibility issue with Windows
+- Fix compatibility issue with Windows and Python 3.11